refactor(product): category, salestate enums 도메인 따로 관리
