+++ src/main/java/com/example/eightyage/domain/product/dto/response/ProductUpdateResponseDto.java
@@ -1,7 +1,7 @@
 package com.example.eightyage.domain.product.dto.response;
 
-import com.example.eightyage.domain.product.entity.Category;
-import com.example.eightyage.domain.product.entity.SaleState;
+import com.example.eightyage.domain.product.category.Category;
+import com.example.eightyage.domain.product.salestate.SaleState;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
