--- src/main/java/com/example/eightyage/domain/product/entity/SaleState.java
@@ -1,6 +1,0 @@
-package com.example.eightyage.domain.product.entity;
-
-public enum SaleState {
-    FOR_SALE,
-    SOLD_OUT
-}
