+++ src/main/java/com/example/eightyage/domain/product/salestate/SaleState.java
@@ -1,0 +1,6 @@
+package com.example.eightyage.domain.product.salestate;
+
+public enum SaleState {
+    FOR_SALE,
+    SOLD_OUT
+}
