+++ src/main/java/com/example/eightyage/domain/product/service/ProductService.java
@@ -3,9 +3,9 @@ package com.example.eightyage.domain.product.service;
 import com.example.eightyage.domain.product.dto.request.ProductSaveRequestDto;
 import com.example.eightyage.domain.product.dto.request.ProductUpdateRequestDto;
 import com.example.eightyage.domain.product.dto.response.*;
-import com.example.eightyage.domain.product.entity.Category;
+import com.example.eightyage.domain.product.category.Category;
 import com.example.eightyage.domain.product.entity.Product;
-import com.example.eightyage.domain.product.entity.SaleState;
+import com.example.eightyage.domain.product.salestate.SaleState;
 import com.example.eightyage.domain.product.repository.ProductRepository;
 import com.example.eightyage.domain.review.entity.Review;
 import com.example.eightyage.domain.review.repository.ReviewRepository;
